+++ src/app.module.ts
@@ -1,6 +1,6 @@
 /** npm imports */
 import { Module } from '@nestjs/common'
-import { ConfigModule, ConfigService } from '@nestjs/config'
+import { ConfigModule } from '@nestjs/config'
 
 /** local imports */
 import appConfig from './config/app.config'
