+++ src/database/types.ts
@@ -1,0 +1,10 @@
+export type WorkoutData = {
+  id: string
+  name: string
+  mode: string
+  equipment: string[]
+  exercises: string[]
+  createdAt: string
+  updatedAt: string
+  trainerTips: string[]
+}
